マークダウンパーサを別ディレクトリへ切り出し、Block・Inlineの処理を別モジュールへ分離
マークダウンパーサの責務が大きくなってきており、公開メソッドも肥大化していた
Block・Inlineで2つに分けることができると考えたため、元のパーサからBlock・Inlineの処理を切り出した
これにより、モジュールの責務が明確になり、テスタビリティも担保できる
